requirements.txt
requests~=2.32.3
pillow~=10.3.0
transformers~=4.41.2
torch~=2.3.1
einops~=0.8.0
timm~=1.0.7
